inference_server/requirements.txt
# Library dependencies for the python code.  You need to install these with
# `pip install -r requirements.txt` before you can run this.
gunicorn
gevent
flask
Pillow
pdf2image
tensorflow
scikit-image
distance
keras